+++ src/SampleApp/libs/dataTestAttributes/appRoles.ts
@@ -1,9 +1,0 @@
-import { roles } from '@/libs/dataTestAttribute/roles'
-import { deepMerge } from '@productive-codebases/toolbox'
-
-export const appRoles = deepMerge([
-  roles,
-  {
-    customContainer: 'customContainer'
-  }
-])