+++ src/index.ts
@@ -1,44 +1,18 @@
 /**
- * Configurator
+ * Configuration
  */
 
-export { configureReactToolbox } from './libs/configureReactToolbox'
-
-/**
- * Logger
- */
-
-export { loggerSetup } from './libs/logger'
-
-/**
- * ForwardProps
- */
-
-export { forwardProps } from './libs/forwardProps'
+export { configure } from './libs/configurator'
 export type {
-  IForwardedProps,
-  ExtractAttributes,
-  ReactHTMLAttributes
-} from './libs/forwardProps/types'
+  IConfigurationParameters,
+  IProviderValue,
+  Configuration
+} from './types'
 
-/**
- * getDataTestAttributeProp
- */
-
-export { roles } from './libs/dataTestAttribute/roles'
-export type {
-  IDataTestAttributeContext,
-  IDataTestAttributeOptions,
-  DataTestAttribute,
-  DataTestAttributeProp
-} from './libs/dataTestAttribute/type'
-
-/**
- * Theme
- */
-
-export { themes } from './styles/themes'
-export { extendTheme } from './libs/theme/extendTheme'
+export {
+  defaultConfiguration,
+  createConfiguration
+} from './libs/configurator/createConfiguration'
 
 /**
  * Components
@@ -47,12 +21,31 @@ export { extendTheme } from './libs/theme/extendTheme'
 export type { IContainerFlexProps } from './components/ContainerFlex/types'
 
 /**
- * Types
+ * Tooling
  */
 
+// named context
+export { createNamedContext } from './libs/namedContext'
+export * from './libs/namedContext/types'
+
+// data-test attributes
 export type {
-  IReactToolboxProviderConfiguration,
-  IReactToolboxProvider as IReactToolboxProviderContext
-} from './types'
+  IDataTestAttributeContext,
+  IDataTestAttributeOptions,
+  DataTestAttribute,
+  DataTestAttributeProp
+} from './libs/dataTestAttribute/type'
+
+// forwardProps
+export { forwardProps } from './libs/forwardProps'
+export type {
+  IForwardedProps,
+  ExtractAttributes,
+  ReactHTMLAttributes
+} from './libs/forwardProps/types'
+
+/**
+ * Types
+ */
 
 export type { WithInnerRef } from './types/reactHelpers'