+++ src/libs/configurator/__tests__/toolboxC.test-fixture.ts
@@ -1,0 +1,21 @@
+/* eslint-disable @typescript-eslint/naming-convention */
+
+import { configure } from '..'
+import { createConfiguration } from '../createConfiguration'
+
+export type ToolboxCProviderConfiguration = typeof toolboxCProviderConfiguration
+
+export const toolboxCProviderConfiguration = createConfiguration({
+  theme: {
+    toolbox: 'toolboxC'
+  },
+  roles: {
+    'toolboxC/name': 'toolboxC/name'
+  }
+})
+
+export const {
+  components: ToolboxCComponents,
+  hooks: toolboxCHooks,
+  helpers: toolboxCHelpers
+} = configure<ToolboxCProviderConfiguration>('contextC')