+++ src/libs/configurator/createConfiguration/portals.ts
@@ -1,0 +1,5 @@
+export const portalNames = {
+  default: 'default'
+}
+
+export type PortalNames = typeof portalNames