+++ src/styles/themes/index.ts
@@ -1,7 +1,0 @@
-import { defaultTheme } from './default'
-
-export type Theme = typeof defaultTheme
-
-export const themes = {
-  default: defaultTheme
-}