+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
+## v0.1.0-beta.32
+
+- Implement and expose a generic NamedContext.
+- Revamp configuration creation and setup.
+
 ## v0.1.0-beta.27
 
-Support nested React contextes.
+- Support named React contextes for nested toolboxes.
 
+## Prior
+
+- Add various components and tooling.